LayoutTests/svg/dom/viewspec-parser-1-expected.txt

This test fuzzes the transform parser with semi-random attribute values and dumps the results of any values that parse successfully.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Loaded: resources/viewspec-target.svg#
Parsed: undefined from: [initial view]

Loaded: resources/viewspec-target.svg#svgView(viewBox(0,0,200,200)))
Parsed: undefined from: svgView(viewBox(0,0,200,200)))

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatio(xMaxYMin%20slice))
Parsed: undefined from: svgView(preserveAspectRatio(xMaxYMin slice))

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatio(xMaxYMin))
Parsed: undefined from: svgView(preserveAspectRatio(xMaxYMin))

Loaded: resources/viewspec-target.svg#svgView(viewBox(1,2,3,4);preserveAspectRatio(xMaxYMin))
Parsed: undefined from: svgView(viewBox(1,2,3,4);preserveAspectRatio(xMaxYMin))

Loaded: resources/viewspec-target.svg#svgView(zoomAndPan(disable))
Parsed: undefined from: svgView(zoomAndPan(disable))

Loaded: resources/viewspec-target.svg#svgView(viewTarget(#myElementId))
Parsed: undefined from: svgView(viewTarget(#myElementId))

Loaded: resources/viewspec-target.svg#svgView
Parsed: undefined from: svgView

Loaded: resources/viewspec-target.svg#svgView(
Parsed: undefined from: svgView(

Loaded: resources/viewspec-target.svg#svgView()
Parsed: undefined from: svgView()

Loaded: resources/viewspec-target.svg#svgView(
Parsed: undefined from: svgView(

Loaded: resources/viewspec-target.svg#svgView(transform(scale(2));;)
Parsed: undefined from: svgView(transform(scale(2));;)

Loaded: resources/viewspec-target.svg#svgView(transform(scale(2));x
Parsed: undefined from: svgView(transform(scale(2));x

Loaded: resources/viewspec-target.svg#svgView(transform(scale(.5)))
Parsed: undefined from: svgView(transform(scale(.5)))

Loaded: resources/viewspec-target.svg#svgView(;transform(scale(.5)))
Parsed: undefined from: svgView(;transform(scale(.5)))

Loaded: resources/viewspec-target.svg#svgView(;;transform(scale(.5)))
Parsed: undefined from: svgView(;;transform(scale(.5)))

Loaded: resources/viewspec-target.svg#svgView(transform(scale(.5));transform(scale(2));transform(scale(2)))
Parsed: undefined from: svgView(transform(scale(.5));transform(scale(2));transform(scale(2)))

Loaded: resources/viewspec-target.svg#svgView(viewBoxx(0,%200)
Parsed: undefined from: svgView(viewBoxx(0, 0)

Loaded: resources/viewspec-target.svg#svgView(viewBo(0,%200)
Parsed: undefined from: svgView(viewBo(0, 0)

Loaded: resources/viewspec-target.svg#svgView(viewBox
Parsed: undefined from: svgView(viewBox

Loaded: resources/viewspec-target.svg#svgView(viewBox
Parsed: undefined from: svgView(viewBox

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatioo(0,%200)
Parsed: undefined from: svgView(preserveAspectRatioo(0, 0)

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRati(0,%200)
Parsed: undefined from: svgView(preserveAspectRati(0, 0)

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatio
Parsed: undefined from: svgView(preserveAspectRatio

Loaded: resources/viewspec-target.svg#svgView(preserveAspectRatio
Parsed: undefined from: svgView(preserveAspectRatio

Loaded: resources/viewspec-target.svg#svgView(transformm(0,%200)
Parsed: undefined from: svgView(transformm(0, 0)

Loaded: resources/viewspec-target.svg#svgView(transfor(0,%200)
Parsed: undefined from: svgView(transfor(0, 0)

Loaded: resources/viewspec-target.svg#svgView(transform
Parsed: undefined from: svgView(transform

Loaded: resources/viewspec-target.svg#svgView(transform
Parsed: undefined from: svgView(transform

Loaded: resources/viewspec-target.svg#svgView(zoomAndPann(0,%200)
Parsed: undefined from: svgView(zoomAndPann(0, 0)

Loaded: resources/viewspec-target.svg#svgView(zoomAndPa(0,%200)
Parsed: undefined from: svgView(zoomAndPa(0, 0)

Loaded: resources/viewspec-target.svg#svgView(zoomAndPan
Parsed: undefined from: svgView(zoomAndPan

Loaded: resources/viewspec-target.svg#svgView(zoomAndPan
Parsed: undefined from: svgView(zoomAndPan

Loaded: resources/viewspec-target.svg#svgView(viewTargett(0,%200)
Parsed: undefined from: svgView(viewTargett(0, 0)

Loaded: resources/viewspec-target.svg#svgView(viewTarge(0,%200)
Parsed: undefined from: svgView(viewTarge(0, 0)

Loaded: resources/viewspec-target.svg#svgView(viewTarget
Parsed: undefined from: svgView(viewTarget

Loaded: resources/viewspec-target.svg#svgView(viewTarget
Parsed: undefined from: svgView(viewTarget

PASS successfullyParsed is true

TEST COMPLETE
